Factor python test functions for EquationsOfState
This required passing the file name into a test helper.
