+++ tests/Unit/PointwiseFunctions/Hydro/EquationsOfState/Test_DarkEnergyFluid.cpp
@@ -39,41 +39,43 @@ SPECTRE_TEST_CASE("Unit.PointwiseFunctions.EquationsOfState.DarkEnergyFluid",
   const DataVector dv_for_size(5);
 
   TestHelpers::EquationsOfState::check(EoS::DarkEnergyFluid<true>(1.0),
-                                       "dark_energy_fluid", d_for_size, 1.0);
+                                       "DarkEnergyFluid", "dark_energy_fluid",
+                                       d_for_size, 1.0);
   TestHelpers::EquationsOfState::check(EoS::DarkEnergyFluid<true>(1.0 / 3.0),
-                                       "dark_energy_fluid", dv_for_size,
-                                       1.0 / 3.0);
+                                       "DarkEnergyFluid", "dark_energy_fluid",
+                                       dv_for_size, 1.0 / 3.0);
   TestHelpers::EquationsOfState::check(EoS::DarkEnergyFluid<true>(1.0),
-                                       "dark_energy_fluid", dv_for_size, 1.0);
+                                       "DarkEnergyFluid", "dark_energy_fluid",
+                                       dv_for_size, 1.0);
   TestHelpers::EquationsOfState::check(EoS::DarkEnergyFluid<true>(1.0 / 3.0),
-                                       "dark_energy_fluid", d_for_size,
-                                       1.0 / 3.0);
+                                       "DarkEnergyFluid", "dark_energy_fluid",
+                                       d_for_size, 1.0 / 3.0);
 
   TestHelpers::EquationsOfState::check(
       TestHelpers::test_creation<
           std::unique_ptr<EoS::EquationOfState<true, 2>>>(
           {"DarkEnergyFluid:\n"
            "  ParameterW: 1.0\n"}),
-      "dark_energy_fluid", d_for_size, 1.0);
+      "DarkEnergyFluid", "dark_energy_fluid", d_for_size, 1.0);
   TestHelpers::EquationsOfState::check(
       TestHelpers::test_creation<
           std::unique_ptr<EoS::EquationOfState<true, 2>>>(
           {"DarkEnergyFluid:\n"
            "  ParameterW: 0.3333333333333333\n"}),
-      "dark_energy_fluid", d_for_size, 1.0 / 3.0);
+      "DarkEnergyFluid", "dark_energy_fluid", d_for_size, 1.0 / 3.0);
 
   TestHelpers::EquationsOfState::check(
       TestHelpers::test_creation<
           std::unique_ptr<EoS::EquationOfState<true, 2>>>(
           {"DarkEnergyFluid:\n"
            "  ParameterW: 1.0\n"}),
-      "dark_energy_fluid", dv_for_size, 1.0);
+      "DarkEnergyFluid", "dark_energy_fluid", dv_for_size, 1.0);
   TestHelpers::EquationsOfState::check(
       TestHelpers::test_creation<
           std::unique_ptr<EoS::EquationOfState<true, 2>>>(
           {"DarkEnergyFluid:\n"
            "  ParameterW: 0.3333333333333333\n"}),
-      "dark_energy_fluid", dv_for_size, 1.0 / 3.0);
+      "DarkEnergyFluid", "dark_energy_fluid", dv_for_size, 1.0 / 3.0);
 
   check_bounds();
 }
