+++ tests/Unit/PointwiseFunctions/Hydro/EquationsOfState/Test_PolytropicFluid.cpp
@@ -70,13 +70,17 @@ SPECTRE_TEST_CASE("Unit.PointwiseFunctions.EquationsOfState.PolytropicFluid",
   const double d_for_size = std::numeric_limits<double>::signaling_NaN();
   const DataVector dv_for_size(5);
   TestHelpers::EquationsOfState::check(EoS::PolytropicFluid<true>{100.0, 2.0},
-                                       "polytropic", d_for_size, 100.0, 2.0);
+                                       "PolytropicFluid", "polytropic",
+                                       d_for_size, 100.0, 2.0);
   TestHelpers::EquationsOfState::check(EoS::PolytropicFluid<true>{134.0, 1.5},
-                                       "polytropic", dv_for_size, 134.0, 1.5);
+                                       "PolytropicFluid", "polytropic",
+                                       dv_for_size, 134.0, 1.5);
   TestHelpers::EquationsOfState::check(EoS::PolytropicFluid<false>{121.0, 1.2},
-                                       "polytropic", d_for_size, 121.0, 1.2);
+                                       "PolytropicFluid", "polytropic",
+                                       d_for_size, 121.0, 1.2);
   TestHelpers::EquationsOfState::check(EoS::PolytropicFluid<false>{117.0, 1.12},
-                                       "polytropic", dv_for_size, 117.0, 1.12);
+                                       "PolytropicFluid", "polytropic",
+                                       dv_for_size, 117.0, 1.12);
 
   TestHelpers::EquationsOfState::check(
       TestHelpers::test_creation<
@@ -84,14 +88,14 @@ SPECTRE_TEST_CASE("Unit.PointwiseFunctions.EquationsOfState.PolytropicFluid",
           {"PolytropicFluid:\n"
            "  PolytropicConstant: 100.0\n"
            "  PolytropicExponent: 2.0\n"}),
-      "polytropic", d_for_size, 100.0, 2.0);
+      "PolytropicFluid", "polytropic", d_for_size, 100.0, 2.0);
   TestHelpers::EquationsOfState::check(
       TestHelpers::test_creation<
           std::unique_ptr<EoS::EquationOfState<true, 1>>>(
           {"PolytropicFluid:\n"
            "  PolytropicConstant: 134.0\n"
            "  PolytropicExponent: 1.5\n"}),
-      "polytropic", dv_for_size, 134.0, 1.5);
+      "PolytropicFluid", "polytropic", dv_for_size, 134.0, 1.5);
 
   TestHelpers::EquationsOfState::check(
       TestHelpers::test_creation<
@@ -99,14 +103,14 @@ SPECTRE_TEST_CASE("Unit.PointwiseFunctions.EquationsOfState.PolytropicFluid",
           {"PolytropicFluid:\n"
            "  PolytropicConstant: 121.0\n"
            "  PolytropicExponent: 1.2\n"}),
-      "polytropic", d_for_size, 121.0, 1.2);
+      "PolytropicFluid", "polytropic", d_for_size, 121.0, 1.2);
   TestHelpers::EquationsOfState::check(
       TestHelpers::test_creation<
           std::unique_ptr<EoS::EquationOfState<false, 1>>>(
           {"PolytropicFluid:\n"
            "  PolytropicConstant: 117.0\n"
            "  PolytropicExponent: 1.12\n"}),
-      "polytropic", dv_for_size, 117.0, 1.12);
+      "PolytropicFluid", "polytropic", dv_for_size, 117.0, 1.12);
 
   check_bounds<true>();
   check_bounds<false>();
